Add test and move the fliipremote code to lib file
