+++ lib/index.js
@@ -1,0 +1,6 @@
+const FlipRemote = require('./flip-remote');
+
+
+module.exports = {
+  FlipRemote
+};
